Made email notification for admin

+++ app/Http/Controllers/AdminController.php
@@ -1,7 +1,6 @@
 <?php
 
 namespace App\Http\Controllers;
-use App\Models\Reviews;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Auth;
 use function Doctrine\Common\Cache\Psr6\get;
@@ -9,6 +8,7 @@ use Illuminate\Routing\Controller as BaseController;
 use App\Models\Roles;
 use App\Models\User;
 use App\Models\Vacancies;
+use App\Models\Reviews;
 use App\Constants;
 use App\Services\Helper;
 use Illuminate\Http\Request;
@@ -69,19 +69,6 @@ class AdminController extends BaseController
             compact('roles', 'candidatesRole', 'companiesRole', 'users', 'userType'));
     }
 
-
-    public function renderVacanciesList() {
-        //TODO left меню вынести в свой компонент blade
-        $roles = $this->roles;
-        $candidatesRole = $this->candidatesRole;
-        $companiesRole = $this->companiesRole;
-
-        $vacancies = Vacancies::all();
-        return view('admin_area.vacancies',
-            compact('roles', 'candidatesRole', 'companiesRole', 'vacancies'));
-    }
-
-
     public function renderUser($id) {
         $roles = $this->roles;
         $candidatesRole = $this->candidatesRole;
@@ -89,28 +76,27 @@ class AdminController extends BaseController
 
         $user = User::find($id);
 
-        if ($user->role_id == Constants::USER_IDS[$candidatesRole]) {
+        if ($user->role_id == Constants::USER_ROLES_IDS[$candidatesRole]) {
             $jobCategories = $this->jobCategories;
             $category = Helper::getTableRow(JobCategories::class, 'ID', $user->CATEGORY_ID);
             return view('admin_area.candidate',
                 compact('user', 'category', 'roles', 'candidatesRole', 'companiesRole', 'jobCategories'));
         }
 
-        if ($user->role_id == Constants::USER_IDS[$companiesRole]) {
+        if ($user->role_id == Constants::USER_ROLES_IDS[$companiesRole]) {
             return view('admin_area.company',
                 compact('user', 'roles', 'candidatesRole', 'companiesRole'));
         }
     }
 
-
     public function updateUserInfo(Request $request)
     {
         $user = User::find($request->id);
 
-        if ($request->role_id == Constants::USER_IDS[$this->companiesRole]) {
+        if ($request->role_id == Constants::USER_ROLES_IDS[$this->companiesRole]) {
             $userList = $this->companiesRole;
             $arrUserFields = User::getCompanyFields();
-        } elseif ($request->role_id == Constants::USER_IDS[$this->candidatesRole]) {
+        } elseif ($request->role_id == Constants::USER_ROLES_IDS[$this->candidatesRole]) {
             $userList = $this->candidatesRole;
             $arrUserFields = User::getCandidateFields();
         }
@@ -125,6 +111,18 @@ class AdminController extends BaseController
         return redirect()->route('admin-users', ['name' => $userList]);
     }
 
+
+
+    public function renderVacanciesList() {
+        //TODO left меню вынести в свой компонент blade
+        $roles = $this->roles;
+        $candidatesRole = $this->candidatesRole;
+        $companiesRole = $this->companiesRole;
+
+        $vacancies = Vacancies::all();
+        return view('admin_area.vacancies',
+            compact('roles', 'candidatesRole', 'companiesRole', 'vacancies'));
+    }
 
     public function renderVacancy($id) {
         $roles = $this->roles;
@@ -161,7 +159,6 @@ class AdminController extends BaseController
 
 
 
-
     public function renderReviewsList() {
         //TODO left меню вынести в свой компонент blade
         $roles = $this->roles;
@@ -169,20 +166,33 @@ class AdminController extends BaseController
         $companiesRole = $this->companiesRole;
 
         $reviews = Reviews::all();
-
-
-//        $review = new Reviews();
-//
-//        $review->NAME = $request->NAME;
-//        $review->REVIEW = $request->REVIEW;
-//        $review->PHOTO = $linkToImage;
-//
-//
-//        $review->save();
-
-
         return view('admin_area.reviews',
             compact('reviews','roles', 'candidatesRole', 'companiesRole'));
+    }
+
+    public function renderReview($id) {
+        $roles = $this->roles;
+        $candidatesRole = $this->candidatesRole;
+        $companiesRole = $this->companiesRole;
+
+        $review = Reviews::find($id);
+
+        return view('admin_area.review',
+            compact('review', 'roles', 'candidatesRole', 'companiesRole'));
+    }
+
+
+    public function updateReview(Request $request)
+    {
+        $review = Reviews::find($request->id);
+        $arrReviewFields = Reviews::getReviewFields();
+
+        foreach ($arrReviewFields as $field) {
+            $review->$field = $request->$field;
+        }
+
+        $review->save();
+        return redirect()->route('admin-reviews');
     }
 
 
@@ -191,14 +201,19 @@ class AdminController extends BaseController
         if (in_array($request->entity, Constants::USER_ENTITIES)) {
             $model = User::class;
             $cachePrefix = 'user_';
-        } elseif ($request->entity = 'vacancy') {
+        } elseif ($request->entity == 'vacancy') {
             $model = Vacancies::class;
             $cachePrefix = 'vacancy_';
+        } elseif ($request->entity == 'review') {
+            $model = Reviews::class;
         }
 
         $entity = $model::find($request->id);
         $entity->delete();
-        $this->cacheService->deleteKeyFromCache($cachePrefix.$request->id);
+
+        if (isset($cachePrefix) && $cachePrefix) {
+            $this->cacheService->deleteKeyFromCache($cachePrefix.$request->id);
+        }
 
         return back();
     }
@@ -207,15 +222,20 @@ class AdminController extends BaseController
         if (in_array($request->entity, Constants::USER_ENTITIES)) {
             $model = User::class;
             $cachePrefix = 'user_';
-        } elseif ($request->entity = 'vacancy') {
+        } elseif ($request->entity == 'vacancy') {
             $model = Vacancies::class;
             $cachePrefix = 'vacancy_';
+        } elseif ($request->entity == 'review') {
+            $model = Reviews::class;
         }
 
         $entity = $model::find($request->id);
         $entity->ACTIVE = ($entity->ACTIVE == 1) ? 0 : 1;
         $entity->save();
-        $this->cacheService->deleteKeyFromCache($cachePrefix.$request->id);
+
+        if (isset($cachePrefix) && $cachePrefix) {
+            $this->cacheService->deleteKeyFromCache($cachePrefix.$request->id);
+        }
 
         return back();
     }
